Add vendorextension for conditional generation of PATH parameters
Added vendorextension for conditional generation of parameters, because it was needed for Spring Data Rest Pagination. Spring Data Rest uses its own paging object which is automatically generated, so you don't need to add it in the OpenApi spec. But because its not in the spec it won't be shown in SwaggerUI, thats why pagination parameters need to be in the spec but won't be generated. As for now it only works with PATH parameters.