Fixed handling of references different from '#/components/...'
This fixes an issue with correct notation of discriminator mapping in oneOf-inheritance structures. According to the specification the discriminator mapping should include a working reference to the target schema - like the oneOf section itself, just without the "$ref" keyword. However, the getSimpleRef() method was not able to resolve such references correctly to the simple name.